requirements.txt
discord-py-slash-commands
discord.py
Pillow
psutil
requests
xlrd
xlutils
xlwt
youtube-dl
mcstatus
icalendar